refactor/ api fetch를 서버 컴포넌트에서 하고 클라이언트 컴포넌트 props로 내려주는 형식으로 수정
